Added error checking

+++ CAB301_Assignment_3/TaskFunctions.cs
@@ -16,6 +16,12 @@ namespace ProjectManagementSystem
         internal static Dictionary<string, Task> s_taskDict = new Dictionary<string, Task>();
         internal static string? s_fileName = "";
 
+        private static void FileError(string message)
+        {
+            Console.WriteLine("! Error: " + message + "\n! Please fix the file and restart the application.\n! Press any key to exit\n...");
+            Console.ReadKey();
+            Environment.Exit(2);
+        }
         public static void LoadTasks(string[] tasksText)
         {
             string[][] taskItemsArray = tasksText.Select(taskString => taskString.Split(new[] { ',' }, 3)).ToArray();
@@ -33,41 +39,74 @@ namespace ProjectManagementSystem
             {
                 Task task = Tasks[i];
 
-                uint timeToCompletion = uint.Parse(taskItemsArray[i][1]); // Add try/catch {|!|}
+                if (!uint.TryParse(taskItemsArray[i][1], out uint timeToCompletion))
+                {
+                    FileError($"the file '{s_fileName}' has invalid time {taskItemsArray[i][1].Trim()} for task {task.ID}.");
+                    
+                }
 
                 // Getting dependencies from text to obj
                 HashSet<Task> dependencies = new HashSet<Task>();
 
-                string[] dependenciesString = taskItemsArray[i][2].Split(",");
-
-                foreach (string dependencyID in dependenciesString)
+                if (taskItemsArray[i].Length == 3)
                 {
-                    s_taskDict.TryGetValue(dependencyID.Trim(), out Task dependency);
-                    dependencies.Add(dependency);
-                }
-                
-                dependencies.Remove(null);
+                    string[] dependenciesString = taskItemsArray[i][2].Split(",");
+
+                    if (dependenciesString[0] != " ")
+                    {
+                        foreach (string dependencyID in dependenciesString)
+                        {
+                            if (!s_taskDict.TryGetValue(dependencyID.Trim(), out Task dependency))
+                            {
+                                FileError($"the file '{s_fileName}' does not have task '{dependencyID.Trim()}', though it is listed as a dependency for task '{task.ID}'");
+                            }
+                            dependencies.Add(dependency);
+                        }
+                    }
+                } 
                 task.Update(timeToCompletion, dependencies);
             }
         }
-        public static void AddTask(string id, uint timeToCompletion, string[] dependenciesIDs)
+        public static void AddTask(string id, uint timeToCompletion, string[]? dependenciesIDs)
         {
             HashSet<Task> dependencies = new HashSet<Task>();
 
-            foreach (string dependencyID in dependenciesIDs)
+            if (dependenciesIDs is not null)
             {
-                s_taskDict.TryGetValue(dependencyID.Trim(), out Task dependency);
-                dependencies.Add(dependency);
+                foreach (string dependencyID in dependenciesIDs)
+                {
+                    if (!s_taskDict.TryGetValue(dependencyID.Trim(), out Task dependency))
+                    {
+                        Console.WriteLine($" ~ Task '{dependencyID.Trim()}' could not be found.\n ~ Press any key to return to the home screen\n...");
+                        Console.ReadKey();
+                        return;
+                    }
+                    dependencies.Add(dependency);
+                }
             }
 
             Task task = new Task(id, timeToCompletion, dependencies);
             Tasks.Add(task);
-            s_taskDict.Add(id, task);
+            try
+            {
+                s_taskDict.Add(id, task);
+            }
+            catch
+            {
+                Console.WriteLine($" ~ Task '{id}' already exists. Task list was not updated.\n ~ Press any key to return to the home screen\n...");
+                Console.ReadKey();
+                return;
+            }
+            
         }
         public static void DeleteTask(string idToDelete)
         {
-            s_taskDict.TryGetValue(idToDelete.Trim(), out Task taskToDelete);
-
+            if (!s_taskDict.TryGetValue(idToDelete.Trim(), out Task taskToDelete))
+            {
+                Console.WriteLine($" ~ Task '{idToDelete}' could not be found.\n ~ Press any key to return to the home screen\n...");
+                Console.ReadKey();
+                return;
+            }
             foreach (Task task in Tasks)
             {
                 task.DeleteDependency(taskToDelete);
@@ -85,7 +124,12 @@ namespace ProjectManagementSystem
         }
         public static void ChangeTime(string id, uint newTime)
         {
-            s_taskDict.TryGetValue(id.Trim(), out Task task);
+            if (!s_taskDict.TryGetValue(id.Trim(), out Task task))
+            {
+                Console.WriteLine($" ~ Task '{id}' could not be found.\n ~ Press any key to return to the home screen\n...");
+                Console.ReadKey();
+                return;
+            }
             task.ChangeTime(newTime);
         }
         public static void Save()
@@ -93,7 +137,7 @@ namespace ProjectManagementSystem
             string[] taskStrings = Tasks.Select(task => task.FormatFileString()).ToArray();
             File.WriteAllLines(s_fileName, taskStrings);
         }
-        public static List<Task> Sequence()
+        public static List<Task>? Sequence()
         {
             List<Task> topOrdering = new List<Task>();
 
@@ -106,7 +150,17 @@ namespace ProjectManagementSystem
                 Task v = tasksClone.Keys.ToList().Find(x => tasksClone.GetValueOrDefault(x).Count == 0);
                 topOrdering.Add(v);
 
-                tasksClone.Remove(v);
+                try
+                {
+                    tasksClone.Remove(v);
+                }
+                catch
+                {
+                    Console.WriteLine($" ~ Tasks have a dependency loop.\n ~ Please fix the tasks and try again.\n ~ Press any key to return to the home screen\n...");
+                    Console.ReadKey();
+                    return null;
+                }
+
                 foreach (HashSet<string> dependencySet in tasksClone.Values)
                 {
                     dependencySet.Remove(v.ID);
@@ -116,10 +170,14 @@ namespace ProjectManagementSystem
         }
         public static void EarliestTimes()
         {
-            List<Task> topOrdering = Sequence();
-            topOrdering.ForEach(x => x.ClearNetwork());
+            List<Task> sequence = Sequence();
+            if (sequence is null)
+            {
+                throw new Exception("Tasks have a dependency loop.");
+            }
+            sequence.ForEach(x => x.ClearNetwork());
 
-            foreach (Task node in topOrdering)
+            foreach (Task node in sequence)
             {
                 if (!node.HasDependencies)
                 {
